pyproject.toml
[tool.poetry]
name = "manim-voiceover"
version = "0.3.7"
description = "Manim plugin for all things voiceover"
authors = ["The Manim Community Developers <contact@manim.community>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "manim_voiceover" }]
repository = "https://github.com/ManimCommunity/manim-voiceover"
homepage = "https://voiceover.manim.community"
documentation = "https://voiceover.manim.community"
keywords = [
    "text-to-speech",
    "tts",
    "voiceover",
    "manim",
    "recording",
    "speech synthesis",
    "math animations",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Sound/Audio :: Capture/Recording",
    "Topic :: Multimedia :: Sound/Audio :: Sound Synthesis",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Natural Language :: English",
]

[tool.poetry.scripts]
manim_translate = 'manim_voiceover.translate.translate:main'
manim_render_translation = 'manim_voiceover.translate.render:main'

[tool.poetry.dependencies]
python = ">=3.8,<4"
manim = "*"
sox = "^1.4.1"
python-dotenv = ">=0.21.0"
mutagen = "^1.46.0"
pydub = "^0.25.1"
azure-cognitiveservices-speech = { version = "^1.24.0", optional = true }
PyAudio = { version = "^0.2.12", optional = true }
gTTS = { version = "^2.2.4", optional = true }
pyttsx3 = { version = "^2.90", optional = true }
openai = { version = "^1.6.1", optional = true }
# torch = { version = "*", optional = true }
# TTS = { version = "*", optional = true }
pynput = { version = "^1.7.6", optional = true }
# deep-translator = { version = "^1.9.2", optional = true }
deepl = { version = "^1.12.0", optional = true }
pip = ">=21.0.1"
openai-whisper = { version = "^20230314", optional = true }
stable-ts = { version = "^2.6.2", optional = true }
python-slugify = "^8.0.1"
elevenlabs = { version = "^0.2.27", optional = true }
replicate = "^1.0.7"

[tool.poetry.extras]
azure = ["azure-cognitiveservices-speech"]
gtts = ["gTTS"]
openai = ["openai"]
pyttsx3 = ["pyttsx3"]
# coqui = ["torch", "TTS"]
coqui = [] # Removed TTS as deps for now
recorder = ["PyAudio", "pynput"]
translate = ["deepl"]
elevenlabs = ["elevenlabs"]
transcribe = ["openai-whisper", "stable-ts"]
all = [
    "azure-cognitiveservices-speech",
    "gTTS",
    "pyttsx3",
    "torch",
    "TTS",
    "PyAudio",
    "pynput",
    "openai",
    "deepl",
    "openai-whisper",
    "stable-ts",
    "elevenlabs",
]

[tool.poetry.group.dev.dependencies]
Pygments = "^2.13.0"
pytest-cov = "^3.0.0"
pytest = "^6.0"
pylint = "^2.12.2"
Sphinx = "^5"
sphinx-copybutton = "^0.4.0"
sphinxext-opengraph = "^0.5.1"
furo = "^2022.9.29"
recommonmark = "^0.7.1"
matplotlib = "^3.3.2"
pre-commit = "^2.11.1"
gitpython = "^3"
pygithub = "^1"
isort = "^5.8.0"
pytest-xdist = "^2.2"
mypy = "^0.931"
types-requests = "^2.25.6"
sphinxcontrib-programoutput = "^0.17"
psutil-wheels = { version = "5.8.0", python = ">=3.10" }
psutil = { version = "^5.8.0", python = "<3.10" }
sphinxcontrib-video = "^0.0.1.dev3"
sphinx-sitemap = "^2.2.1"


[tool.poetry.plugins."manim.plugins"]
"manim_voiceover" = "manim_voiceover"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
